backend/requirements.txt
# Core
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
orjson
websockets

# ASR
openai-whisper
torch
torchaudio
transformers